Stage 5 -- Infinite loop on test #3

+++ JSON Database/task/src/server/database/Database.java
@@ -1,27 +1,68 @@
 package server.database;
 
+import com.google.gson.Gson;
 import server.commands.Request;
 import server.commands.Response;
 
+import java.io.*;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.concurrent.locks.Lock;
+import java.util.concurrent.locks.ReadWriteLock;
+import java.util.concurrent.locks.ReentrantReadWriteLock;
 
 public class Database {
-    static final String ERROR = "Error";
+    static final String ERROR = "ERROR";
     static final String NO_SUCH_KEY = "No such key";
     static final String OK = "OK";
+    static final String PATH = "/Users/ceramesupersound/IntelliJ Projects/JSON Database/JSON Database/task/src/server/data/db.json";
+    ReadWriteLock lock;
+    Lock readLock, writeLock;
 
-    private Map<String, String> database = new HashMap<>();
 
-    public Response getRecord(String key) {
+    public Database() {
+        lock = new ReentrantReadWriteLock();
+        readLock = lock.readLock();
+        writeLock = lock.writeLock();
+        init();
+    }
+
+    private void init() {
+        try {
+            writeLock.lock();
+            FileWriter writer = new FileWriter(PATH);
+            writer.write("{}");
+            writer.close();
+
+            writeLock.unlock();
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+    }
+
+    public Response getRecord(Request request) {
         Response response = new Response();
-        String value = database.get(key);
-        if (value == null) {
+        String value;
+
+        try {
+            readLock.lock();
+
+            FileReader reader = new FileReader(PATH);
+            Map<String, String> database = new Gson().fromJson(reader, Map.class);
+            reader.close();
+            response.setValue(database.get(request.getKey()));
+
+            readLock.unlock();
+
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+
+        if (response.getValue() == null) {
             response.setResponse(ERROR);
             response.setReason(NO_SUCH_KEY);
         } else {
             response.setResponse(OK);
-            response.setValue(value);
         }
         return response;
 
@@ -29,11 +70,28 @@ public class Database {
 
     public Response setRecord(Request request) {
         Response response = new Response();
-        database.put(request.getKey(), request.getValue());
 
-        response.setResponse(OK);
-        response.setValue(null);
-        response.setReason(null);
+        try {
+            writeLock.lock();
+
+            FileReader reader = new FileReader(PATH);
+            Map<String, String> database = new Gson().fromJson(reader, Map.class);
+            reader.close();
+
+            database.put(request.getKey(), request.getValue());
+            FileWriter writer = new FileWriter(PATH);
+            writer.write(new Gson().toJson(database));
+            writer.close();
+            response.setResponse(OK);
+            response.setReason(null);
+
+
+            writeLock.unlock();
+
+        } catch (Exception ignored) {
+
+        }
+
 
         return response;
 
@@ -41,14 +99,24 @@ public class Database {
 
     public Response deleteRecord(Request request) {
         Response response = new Response();
-        String deletedRecord = database.remove(request.getKey());
+        try {
+            writeLock.lock();
+            FileReader reader = new FileReader(PATH);
+            Map<String, String> database = new Gson().fromJson(reader, Map.class);
+            reader.close();
 
-        if (deletedRecord == null) {
-            response.setResponse(ERROR);
-            response.setReason(NO_SUCH_KEY);
-        } else {
-            response.setResponse(OK);
+            if (database.containsKey(request.getKey())) {
+                database.remove(request.getKey());
+                response.setResponse(OK);
+
+            } else {
+                response.setResponse(ERROR);
+                response.setReason(NO_SUCH_KEY);
+            }
+        } catch (Exception e) {
+            e.printStackTrace();
         }
+
         return response;
     }
